Create Ex01 - my_first_variable_integer.c
First part of coding is to create variable. Let's get started with an integer variable.
What does it mean "integer variable?"
In most languages you have "types", a good comparaison is a letter is different from a number.
In a computer everything is numbers (0 and 1). But we, as human, interact with letter (and words) to make it usuable there are "convention": a letter is a number and one of this table of convertion is: ASCII (you should google man ascii).