Create Ex02 - my_first_variable_char.c
The first part of coding is to create a variable. Let's get started with an character variable.
What does a "character variable” mean?
In most languages, you have "types". A good comparison is that a letter is different from a number.
In a computer, everything is numbered (0 and 1). But we, as humans, interact with a letter (and words) to make it useable there is "convention": a letter is a number and one of this table of conversion is: ASCII (you should google man ASCII)