Create Ex14 - 	my_isalpha.c
Function isalpha() takes a single argument in the form of an integer and returns an integer value. Even though isalpha() takes an integer as an argument, a character is passed to isalpha() function. Internally, the character is converted into the integer value corresponding to its ASCII value when passed.
Created a my_isalpha function that reproduces the behavior of isalpha() function. It returns 1 if the character sent as an argument is a letter (A to Z or a to z). It returns 0 otherwise.