Create Ex18 - my_isspace.c
isspace() checks for white-space characters.  In the "C" and "POSIX" locales, these are: space, form-feed ('\f'), newline ('\n'), carriage return ('\r'), horizontal tab ('\t'), and vertical tab ('\v').
Created a my_isspace function that reproduces the behavior of isspace() function. It returns 1 if the character sent as an argument is a space. It returns 0 otherwise.