Create Ex08 - my_string_index.c
Created a my_string_index function that takes 2 parameters (haystack and needle) and locates the first occurrence of the character needle in the string haystack and returns the position. The objective was to build a loop and having an if statement when it matches the right character and returns its position.