Update Ex07 - my_first_struct.c
Structure in C are similar to the concept of package. You can have multiple objects inside your package and with a struct, you can have multiple variables in it. 
A struct is a type. I'm defining a struct of type struct s_box which will contain variables. 
A struct has a special behavior if it's a pointer on a struct or not. If it's a pointer on a struct, in order to reach the variable I use '->'.
In this exercise, I received a struct integer_array and I had to print its size and its content following this format: size\narray[0]\narray[1]...
If the size is 0, just print 0\n.