Create Ex08 - my_is_sort.c
Created a function that tells us if an array is sorted or not. What is sorted? :-)
This function takes an integer array as a parameter (input) and returns a boolean (true/false).
My function returns true if the integer array is sorted in either ASC(ascending) or DESC(descending) order.
My function returns false if the integer array is not sorted.