Create Ex00 - my_count_files
Back to do some shell.
The objective is to be more expert with find and discover -type file option, globing (* replace any characters...) and --exec to be able to run another command. (for example basename :D) Finally, pipe (|) lets you chain command.
Wrote a command line that counts and displays the number of regular files and directories in the current directory and all its sub-directories. It should include ".", the starting directory.