Create Ex01 - my_recursive_pow.c
Created an iterated function that returns the value of a power applied to a number. A power lower than 0 returns 0. Overflows didn't have to be handled.
The first parameter is the number, the second parameter is the power. I had to use the recursive method to perform this exercise. While / for / any loop were forbidden.