Create Ex03 - rcapitalize.c
Wrote a function that takes one string and, capitalizes the last character of each word in uppercase and the rest in lowercase. A word is a section of string delimited by spaces/tabs or the start/end of the string. If a word has a single letter, it must be capitalized. A letter is a character in the set [a-zA-Z].