Create Ex00 - hidenp.c
Wrote a function named hidenp that takes two strings and returns 1 if the first string is hidden in the second one, otherwise returns 0 followed by a newline.
Let s1 and s2 be strings. We say that s1 is hidden in s2 if it's possible to find each character from s1 in s2, in the same order as they appear in s1.
Also, the empty string is hidden in any string.