Create Ex01 - my_split.c
Created a function that splits a string of characters depending on a separator. (sounds crazy....keep reading!). The second argument is a unique character separator.
The function should return an array which contains a string wrapped between separator. There cannot be any empty strings in your array. 
The string given as an argument of the function won't be modifiable.
My logic sequence: 
0# count number of words inside a string
1# alloc the struct
2# alloc the array of pointer inside the struct
3# loop throught the string to extract words
4# malloc for the extracted words
5# "store" them inside the array of the string.
6# set size inside the struct :D